add basic application information to LowBeams
This commit adds basic application information, including title, version, and icon image, to the main application class. The versioning system is explained in detail in the adjacent code comments. The NetBeans project properties file is also update with the application version, and should continue to be for future commits. Finally, this commit also adds a small JavaFX display of the added application information. This code can be removed, as explained in the comments, when actual application logic is implemented, but for now, it serves to demonstrate what has been added.

Fixed Issues:

Closes #4: LowBeams class needs basic application information