allow user to configure overlay color
This commit adds the ability for users to change the color and opacity of the overlay from the preferences menu. It also improves some null value checking in property change listeners. Of note is the fact that the color preferences are implemented as separate base color and opacity controls. This was primarily to avoid issues arising the custom color dialog window created by the default JavaFX color picker control, which would compete with the overlay for focus.

Fixed Issues:

Closes #12: overlay lacks bindable color option
Closes #21: property change listeners should check for null new values