add option to select target screen
This commit primarily adds the ability for users to select the target screen for the overlay, automatically refitting the overlay stage to the selected target screen. Note that this option has not yet been tested on an actual multi-monitor hardware configuration, so it is unconfirmed if it works properly, but at least on a single-display system, the relevant components seem to be working properly.

Fixed Issues:

Closes #15: need a way to specify target screen for an overlay
Closes #25: setting overlays to full screen is redundant
Closes #26: application version string has not been updated recently