general project maintenance
Before, the project did not explicitly specify the latest JDK 8. It also violated the JAR naming conventions (which would have been a problem for native packaging if it weren't for the fact that NetBeans support for it seems horribly buggy and I gave up). This commit also clears a codebase attribute security warning by setting the attribute in the project properties file.

Fixed Issues:

#43: update project to latest JDK 8